# backend/src/Aido.Application/UseCases/TodoItems/Commands/UpdateTodoItem/TodoItemResponse.cs
using Aido.Core;

namespace Aido.Application.UseCases.TodoItems.Commands.UpdateTodoItem;

public class TodoItemResponse
{
    public Guid Id { get; set; }
    public string Title { get; set; } = string.Empty;
    public string Description { get; set; } = string.Empty;
    public bool IsCompleted { get; set; }
    public DateTime CreatedAt { get; set; }
    public DateTime UpdatedAt { get; set; }

    public static TodoItemResponse FromDomain(TodoItem todoItem)
    {
        return new TodoItemResponse
        {
            Id = todoItem.Id,
            Title = todoItem.Title,
            Description = todoItem.Description,
            IsCompleted = todoItem.IsCompleted,
            CreatedAt = todoItem.CreatedAt,
            UpdatedAt = todoItem.UpdatedAt
        };
    }
}